refactor: move solver target artifacts to arco-targets
Move solver-facing linear compile output types into the arco-targets crate and re-export them through existing KDL artifact seams for compatibility. Update solver/export consumers to import the shared target types directly where appropriate.

Resolves #228
