+++ crates/arco-cli/Cargo.toml
@@ -37,6 +37,7 @@ arco-xpress = { workspace = true, optional = true }
 arco-scip = { workspace = true }
 arco-export = { workspace = true }
 arco-solver = { workspace = true }
+arco-targets = { workspace = true }
 clap = { version = "4", features = ["derive"] }
 miette = { version = "7", features = ["fancy"] }
 serde = { version = "1.0", features = ["derive"] }
