+++ crates/arco-cli/src/debug_shell.rs
@@ -1,5 +1,5 @@
-use arco_kdl::artifacts::{AlgebraicProblem, ConstraintSense, ObjectiveSense, VariableKind};
 use arco_kdl::pipeline::compile_file;
+use arco_targets::{AlgebraicProblem, ConstraintSense, ObjectiveSense, VariableKind};
 use miette::{IntoDiagnostic, Result, miette};
 use std::collections::BTreeMap;
 use std::fs::{File, remove_file};
@@ -314,7 +314,7 @@ impl Drop for DebugBootstrapScript {
 #[cfg(test)]
 mod tests {
     use crate::debug_shell::{PythonModelData, build_ipython_script, build_python_model_data};
-    use arco_kdl::artifacts::{
+    use arco_targets::{
         AlgebraicProblem, ConstraintSense, LinearConstraint, LinearObjective, LinearTerm,
         ObjectiveSense, VariableInstance, VariableKind,
     };
