+++ crates/arco-export/Cargo.toml
@@ -14,6 +14,6 @@ publish = false
 workspace = true
 
 [dependencies]
-arco-kdl = { workspace = true }
+arco-targets = { workspace = true }
 miette = { version = "7", features = ["fancy"] }
 thiserror = "2.0"
