+++ crates/arco-kdl/Cargo.toml
@@ -15,6 +15,7 @@ workspace = true
 
 [dependencies]
 arco-core = { workspace = true }
+arco-targets = { workspace = true }
 arco-validate = { workspace = true }
 csv = "1.4"
 kdl = "6.5"
