+++ crates/arco-scip/Cargo.toml
@@ -17,6 +17,7 @@ workspace = true
 arco-export = { workspace = true }
 arco-kdl = { workspace = true }
 arco-solver = { workspace = true }
+arco-targets = { workspace = true }
 miette = { version = "7", features = ["fancy"] }
 thiserror = "2.0"
 tracing = { workspace = true }
