+++ crates/arco-targets/Cargo.toml
@@ -14,3 +14,5 @@ publish = false
 workspace = true
 
 [dependencies]
+arco-core = { workspace = true }
+serde = { version = "1.0", features = ["derive"] }
